Se agrega el modulo Usuarios, contemplando el servicio de usuarios, componenete usuarioList y formulario de creación usuarioForm, se valida funcionamiento de la integración de endpoints de usuarios.  Se realiza un cambio en el modelo Usuarios, para que DNI sea tipo BIGINT y asi aceptar más de 13 digitos, se cambia la validación del DNI ya que no estaba encontrando correctamente la información de la validación, adicional tambien se corrige la funcionalidad en login, para que ahora valide tambien si el usuario esta activo para que permita ingresar al sistema, ya que no lo estaba realizando

+++ BackEnd/app/models.py
@@ -21,7 +21,7 @@ class Usuario(db.Model):
     usuario_id = db.Column(db.Integer, primary_key=True, autoincrement=True)
     nombre = db.Column(db.String(100), nullable=False)
     apellidos = db.Column(db.String(100), nullable=False)
-    dni = db.Column(db.Integer, unique=True, nullable=False)
+    dni = db.Column(db.BIGINT, unique=True, nullable=False)
     correo = db.Column(db.String(100), unique=True)
     telefono = db.Column(db.Integer)
     password_hash = db.Column(db.String(255), nullable=False)
